[TP]: Add missing message length for frame chunk callback Also changed the message length/indexes in CANMessage from `size_t` to `uint32_t` as it can't exceed the maximum message length of ETP: 117440505

+++ isobus/include/isobus/isobus/can_message.hpp
@@ -45,7 +45,7 @@ namespace isobus
 		/// with extended transport protocol is restricted by the extended data packet offset (3 bytes).
 		/// This yields a maximum message size of (2^24-1 packets) x (7 bytes/packet) = 117440505 bytes
 		/// @returns The maximum length of any CAN message as defined by ETP in ISO11783
-		static const std::size_t ABSOLUTE_MAX_MESSAGE_LENGTH = 117440505;
+		static const std::uint32_t ABSOLUTE_MAX_MESSAGE_LENGTH = 117440505;
 
 		/// @brief Constructor for a CAN message
 		/// @param[in] CANPort The can channel index the message uses
@@ -64,7 +64,7 @@ namespace isobus
 
 		/// @brief Returns the length of the data in the CAN message
 		/// @returns The message data payload length
-		virtual std::size_t get_data_length() const;
+		virtual std::uint32_t get_data_length() const;
 
 		/// @brief Gets the source control function that the message is from
 		/// @returns The source control function that the message is from
@@ -85,16 +85,16 @@ namespace isobus
 		/// @brief Sets the message data to the value supplied. Creates a copy.
 		/// @param[in] dataBuffer The data payload
 		/// @param[in] length the length of the data payload in bytes
-		void set_data(const std::uint8_t *dataBuffer, std::size_t length);
+		void set_data(const std::uint8_t *dataBuffer, std::uint32_t length);
 
 		/// @brief Sets one byte of data in the message data payload
 		/// @param[in] dataByte One byte of data
 		/// @param[in] insertPosition The position in the message at which to insert the data byte
-		void set_data(std::uint8_t dataByte, const std::size_t insertPosition);
+		void set_data(std::uint8_t dataByte, const std::uint32_t insertPosition);
 
 		/// @brief Sets the size of the data payload
 		/// @param[in] length The desired length of the data payload
-		void set_data_size(std::size_t length);
+		void set_data_size(std::uint32_t length);
 
 		/// @brief Sets the source control function for the message
 		/// @param[in] value The source control function
@@ -113,7 +113,7 @@ namespace isobus
 		/// @details This function will return the byte at the specified index in the buffer.
 		/// @param[in] index The index to get the byte from
 		/// @return The 8-bit unsigned byte
-		std::uint8_t get_uint8_at(const std::size_t index) const;
+		std::uint8_t get_uint8_at(const std::uint32_t index) const;
 
 		/// @brief Get a 16-bit unsigned integer from the buffer at a specific index.
 		/// A 16-bit unsigned integer can hold a value between 0 and 65535.
@@ -121,7 +121,7 @@ namespace isobus
 		/// @param[in] index The index to get the 16-bit unsigned integer from
 		/// @param[in] format The byte format to use when reading the integer
 		/// @return The 16-bit unsigned integer
-		std::uint16_t get_uint16_at(const std::size_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
+		std::uint16_t get_uint16_at(const std::uint32_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
 
 		/// @brief Get a right-aligned 24-bit integer from the buffer (returned as a uint32_t) at a specific index.
 		/// A 24-bit number can hold a value between 0 and 16,777,215.
@@ -129,7 +129,7 @@ namespace isobus
 		/// @param[in] index The index to get the 24-bit unsigned integer from
 		/// @param[in] format The byte format to use when reading the integer
 		/// @return The 24-bit unsigned integer, right aligned into a uint32_t
-		std::uint32_t get_uint24_at(const std::size_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
+		std::uint32_t get_uint24_at(const std::uint32_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
 
 		/// @brief Get a 32-bit unsigned integer from the buffer at a specific index.
 		/// A 32-bit unsigned integer can hold a value between 0 and 4294967295.
@@ -137,7 +137,7 @@ namespace isobus
 		/// @param[in] index The index to get the 32-bit unsigned integer from
 		/// @param[in] format The byte format to use when reading the integer
 		/// @return The 32-bit unsigned integer
-		std::uint32_t get_uint32_at(const std::size_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
+		std::uint32_t get_uint32_at(const std::uint32_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
 
 		/// @brief Get a 64-bit unsigned integer from the buffer at a specific index.
 		/// A 64-bit unsigned integer can hold a value between 0 and 18446744073709551615.
@@ -145,7 +145,7 @@ namespace isobus
 		/// @param[in] index The index to get the 64-bit unsigned integer from
 		/// @param[in] format The byte format to use when reading the integer
 		/// @return The 64-bit unsigned integer
-		std::uint64_t get_uint64_at(const std::size_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
+		std::uint64_t get_uint64_at(const std::uint32_t index, const ByteFormat format = ByteFormat::LittleEndian) const;
 
 		/// @brief Get a bit-boolean from the buffer at a specific index.
 		/// @details This function will return whether the bit(s) at the specified index in the buffer is/are (all) equal to 1.
@@ -153,7 +153,7 @@ namespace isobus
 		/// @param[in] bitIndex The bit index to start reading the boolean from, ranging from 0 to 7
 		/// @param[in] length The number of bits to read, maximum of (8 - bitIndex)
 		/// @return True if (all) the bit(s) are set, false otherwise
-		bool get_bool_at(const std::size_t byteIndex, const std::uint8_t bitIndex, const std::uint8_t length = 1) const;
+		bool get_bool_at(const std::uint32_t byteIndex, const std::uint8_t bitIndex, const std::uint8_t length = 1) const;
 
 	private:
 		Type messageType = Type::Receive; ///< The internal message type associated with the message
